Wall generation is working. Just need fix doors not exactly lining up with the walls

+++ src/dungeongen/direction.rs
@@ -1,6 +1,6 @@
 use std::slice::Iter;
 
-#[derive(PartialEq, Debug, Clone)]
+#[derive(PartialEq, Debug, Clone, Copy)]
 pub enum Direction {
   North,
   NorthEast,
@@ -37,22 +37,22 @@ impl Direction {
     &DIRECTIONS
   }
 
-  pub fn to_tup(&self) -> (i32, i32) {
+  pub fn to_tup(&self) -> (f32, f32) {
     match *self {
-      Direction::North => (0, 1),
-      Direction::NorthEast => (1, 1),
-      Direction::East => (1, 0),
-      Direction::SouthEast => (1, -1),
-      Direction::South => (0, -1),
-      Direction::SouthWest => (-1, -1),
-      Direction::West => (-1, 0),
-      Direction::NorthWest => (-1, 1),
+      Direction::North => (0.0, 1.0),
+      Direction::NorthEast => (1.0, 1.0),
+      Direction::East => (1.0, 0.0),
+      Direction::SouthEast => (1.0, -1.0),
+      Direction::South => (0.0, -1.0),
+      Direction::SouthWest => (-1.0, -1.0),
+      Direction::West => (-1.0, 0.0),
+      Direction::NorthWest => (-1.0, 1.0),
     }
   }
 
   pub fn dir_from_tup(&self, other: (i32, i32)) -> (i32, i32) {
     let modifier = self.to_tup();
-    (other.0 + modifier.0, other.1 + modifier.1)
+    (other.0 + modifier.0 as i32, other.1 + modifier.1 as i32)
   }
 
   pub fn opposite(&self) -> Direction {
